requirements.txt
pandas
requests
Pillow
pytesseract
